+++ mathics/builtin/strings.py
@@ -11,7 +11,7 @@ from sys import version_info
 import unicodedata
 from binascii import hexlify, unhexlify
 from heapq import heappush, heappop
-from typing import Callable
+from typing import Any, Callable, List
 
 from mathics.version import __version__  # noqa used in loading to check consistency.
 from mathics.builtin.base import BinaryOperator, Builtin, Test, Predefined
@@ -21,6 +21,7 @@ from mathics.core.expression import (
     SymbolFailed,
     SymbolFalse,
     SymbolTrue,
+    SymbolList,
     String,
     Integer,
     from_python,
@@ -39,6 +40,36 @@ _regex_longest = {
 _regex_shortest = {
     "+": "+?",
     "*": "*?",
+}
+
+
+alphabet_descriptions = {
+    "English": {
+        "Lowercase": "abcdefghijklmnopqrstuvwxyz",
+        "Uppercase": "ABCDEFGHIJKLMNOPQRSTUVWXYZ",
+    },
+    "Spanish": {
+        "Lowercase": "abcdefghijklmnñopqrstuvwxyz",
+        "Uppercase": "ABCDEFGHIJKLMNÑOPQRSTUVWXYZ",
+    },
+    "Greek": {
+        "Lowercase": "αβγδεζηθικλμνξοπρστυφχψω",
+        "Uppercase": "ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ",
+    },
+    "Cyrillic": {
+        "Lowercase": "абвгґдђѓеёєжзѕиіїйјклљмнњопрстћќуўфхцчџшщъыьэюя",
+        "Uppercase": "АБВГҐДЂЃЕЁЄЖЗЅИІЇЙЈКЛЉМНЊОПРСТЋЌУЎФХЦЧЏШЩЪЫЬЭЮЯ",
+    },
+}
+
+alphabet_alias={
+    "English": "English",
+    "French": "English",
+    "German": "English",
+    "Spanish": "Spanish",
+    "Greek": "Greek",
+    "Cyrillic": "Cyrillic",
+    "Russian": "Cyrillic",
 }
 
 
@@ -639,6 +670,164 @@ class LetterCharacter(Builtin):
     """
 
 
+# FIXME: Generalize string.lower() and ord()
+def letter_number(chars: List[str], start_ord) -> List["Integer"]:
+    # Note caller has verified that everything isalpha() and
+    # each char has length 1.
+    return [Integer(ord(char.lower()) - start_ord) for char in chars]
+
+
+class Alphabet(Builtin):
+    """
+     <dl>
+      <dt>'Alphabet'[]
+      <dd>gives the list of lowercase letters a-z in the English alphabet .
+
+      <dt>'Alphabet[$type$]'
+      <dd> gives the alphabet for the language or class $type$.
+    </dl>
+
+    >> Alphabet[]
+     = {a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z}
+    >> Alphabet["German"]
+     = {a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z}
+
+    """
+    messages = {
+        "nalph": "The alphabet `` is not known or not available.",
+    }
+
+    rules = {
+        "Alphabet[]": """Alphabet["English"]""",
+    }
+
+    def apply(self, alpha, evaluation):
+        """Alphabet[alpha_String]"""
+        alphakey = alpha.get_string_value()
+        alphakey = alphabet_alias[alphakey]
+        if alphakey is None:
+            evaluation.message("Alphabet", "nalph", alpha)            
+            return
+        alphabet = alphabet_descriptions.get(alphakey, None)
+        if alphabet is None:
+            evaluation.message("Alphabet", "nalph", alpha)
+            return
+        return Expression(SymbolList, *[String(c) for c in alphabet["Lowercase"]])
+
+
+class LetterNumber(Builtin):
+    """
+    <dl>
+      <dt>'LetterNumber'[$c$]
+      <dd>returns the position of the character $c$ in the English alphabet.
+
+      <dt>'LetterNumber["string"]'
+      <dd>returns a list of the positions of characters in string.
+      <dt>'LetterNumber["string", $alpha$]'
+      <dd>returns a list of the positions of characters in string, regarding the alphabet $alpha$.
+    </dl>
+
+    >> LetterNumber["b"]
+     = 2
+
+    LetterNumber also works with uppercase characters
+    >> LetterNumber["B"]
+     = 2
+
+    >> LetterNumber["ss2!"]
+     = {19, 19, 0, 0}
+
+    Get positions of each of the letters in a string:
+    >> LetterNumber[Characters["Peccary"]]
+    = {16, 5, 3, 3, 1, 18, 25}
+
+    >> LetterNumber[{"P", "Pe", "P1", "eck"}]
+    = {16, {16, 5}, {16, 0}, {5, 3, 11}}
+
+    #> LetterNumber[4]
+     : The argument 4 is not a string.
+     = LetterNumber[4]
+
+    >> LetterNumber["\[Beta]", "Greek"]
+     = 2
+    
+    """
+    # FIXME: put the right unicode characters in a way that the
+    # following test works...
+    """
+    # #> LetterNumber["\[CapitalBeta]", "Greek"]
+    #  = 2
+
+    """
+    messages = {
+        "nalph": "The alphabet `` is not known or not available.",
+        "nas": ("The argument `1` is not a string."),
+    }
+
+    def apply_alpha_str(self, chars: List[Any], alpha: String, evaluation):
+        "LetterNumber[chars_, alpha_String]"
+        alphakey = alpha.get_string_value()
+        alphakey = alphabet_alias.get(alphakey, None)
+        if alphakey is None:
+            evaluation.message("LetterNumber", "nalph", alpha)
+            return
+        if alphakey == "English":
+            return self.apply(chars, evaluation)
+        alphabet = alphabet_descriptions.get(alphakey, None)
+        if alphabet is None:
+            evaluation.message("LetterNumber", "nalph", alpha)
+            return
+        # TODO: handle Uppercase
+        if isinstance(chars, String):
+            py_chars = chars.get_string_value()
+            if len(py_chars) == 1:
+                # FIXME generalize ord("a")
+                res = alphabet["Lowercase"].find(py_chars) + 1
+                if res == -1:
+                    res = alphabet["Uppercase"].find(py_chars) + 1
+                return Integer(res)
+            else:
+                r = []
+                for c in py_chars:
+                    cp = alphabet["Lowercase"].find(c) + 1
+                    if cp == -1:
+                        cp = alphabet["Uppercase"].find(c) + 1
+                    r.append(cp)
+                return Expression(SymbolList, *r)
+        elif chars.has_form("List", 1, None):
+            result = []
+            for leaf in chars.leaves:
+                result.append(self.apply_alpha_str(leaf, alpha, evaluation))
+            return Expression(SymbolList, *result)
+        else:
+            return evaluation.message(self.__class__.__name__, "nas", chars)
+        return None
+
+    def apply(self, chars: List[Any], evaluation):
+        "LetterNumber[chars_]"
+
+        start_ord = ord("a") - 1
+        if isinstance(chars, String):
+            py_chars = chars.get_string_value()
+            if len(py_chars) == 1:
+                # FIXME generalize ord("a")
+                return letter_number([py_chars[0]], start_ord)[0]
+            else:
+                r = [
+                    letter_number(c, start_ord)[0] if c.isalpha() else 0
+                    for c in py_chars
+                ]
+                return Expression(SymbolList, *r)
+        elif chars.has_form("List", 1, None):
+            result = []
+            for leaf in chars.leaves:
+                result.append(self.apply(leaf, evaluation))
+            return Expression(SymbolList, *result)
+        else:
+            return evaluation.message(self.__class__.__name__, "nas", chars)
+        return None
+
+
 class HexidecimalCharacter(Builtin):
     """
     <dl>
@@ -848,7 +1037,7 @@ class StringJoin(BinaryOperator):
         "StringJoin[items___]"
 
         result = ""
-        items = items.flatten(Symbol("List"))
+        items = items.flatten(SymbolList)
         if items.get_head_name() == "System`List":
             items = items.leaves
         else:
@@ -933,7 +1122,7 @@ class StringSplit(Builtin):
 
         if string.get_head_name() == "System`List":
             leaves = [self.apply(s, patt, evaluation, options) for s in string._leaves]
-            return Expression("List", *leaves)
+            return Expression(SymbolList, *leaves)
 
         py_string = string.get_string_value()
 
@@ -961,7 +1150,7 @@ class StringSplit(Builtin):
         for re_patt in re_patts:
             result = [t for s in result for t in mathics_split(re_patt, s, flags=flags)]
 
-        return string_list("List", [String(x) for x in result if x != ""], evaluation)
+        return string_list(SymbolList, [String(x) for x in result if x != ""], evaluation)
 
 
 class StringPosition(Builtin):
@@ -1092,7 +1281,7 @@ class StringPosition(Builtin):
                 self.do_apply(py_string, compiled_patts, py_n, overlap)
                 for py_string in py_strings
             ]
-            return Expression("List", *results)
+            return Expression(SymbolList, *results)
         else:
             py_string = string.get_string_value()
             if py_string is None:
@@ -1362,6 +1551,7 @@ class StringReverse(Builtin):
       >> StringReverse["live"]
        = evil
     """
+
     def apply(self, string, evaluation):
         "StringReverse[string_String]"
         return String(string.get_string_value()[::-1])
@@ -1426,7 +1616,7 @@ class StringCases(_StringFind):
                 else:
                     yield _evaluate_match(form, match, evaluation)
 
-        return Expression("List", *list(cases()))
+        return Expression(SymbolList, *list(cases()))
 
     def apply(self, string, rule, n, evaluation, options):
         "%(name)s[string_, rule_, OptionsPattern[%(name)s], n_:System`Private`Null]"
@@ -1510,7 +1700,7 @@ class Characters(Builtin):
     def apply(self, string, evaluation):
         "Characters[string_String]"
 
-        return Expression("List", *(String(c) for c in string.value))
+        return Expression(SymbolList, *(String(c) for c in string.value))
 
 
 class CharacterRange(Builtin):
@@ -1886,7 +2076,7 @@ class ToCharacterCode(Builtin):
         if encoding == "Unicode":
 
             def convert(s):
-                return Expression("List", *[Integer(ord(code)) for code in s])
+                return Expression(SymbolList, *[Integer(ord(code)) for code in s])
 
         else:
             py_encoding = to_python_encoding(encoding)
@@ -1900,7 +2090,7 @@ class ToCharacterCode(Builtin):
                 )
 
         if isinstance(string, list):
-            return Expression("List", *[convert(substring) for substring in string])
+            return Expression(SymbolList, *[convert(substring) for substring in string])
         elif isinstance(string, str):
             return convert(string)
 
@@ -2010,7 +2200,7 @@ class FromCharacterCode(Builtin):
                         evaluation.message(
                             "FromCharacterCode",
                             "notunicode",
-                            Expression("List", *l),
+                            Expression(SymbolList, *l),
                             Integer(i + 1),
                         )
                         raise _InvalidCodepointError
@@ -2035,7 +2225,7 @@ class FromCharacterCode(Builtin):
                         else:
                             stringi = convert_codepoint_list([leaf])
                         list_of_strings.append(String(stringi))
-                    return Expression("List", *list_of_strings)
+                    return Expression(SymbolList, *list_of_strings)
                 else:
                     return String(convert_codepoint_list(n.get_leaves()))
             else:
@@ -2450,9 +2640,7 @@ def _damerau_levenshtein(s1, s2, sameQ: Callable[..., bool]):
     return d_prev[-1]
 
 
-def _levenshtein_like_or_border_cases(
-    s1, s2, sameQ: Callable[..., bool], compute
-):
+def _levenshtein_like_or_border_cases(s1, s2, sameQ: Callable[..., bool], compute):
     if len(s1) == len(s2) and all(sameQ(c1, c2) for c1, c2 in zip(s1, s2)):
         return 0
 
@@ -2853,7 +3041,7 @@ def _pattern_search(name, string, patt, evaluation, options, matched):
             return evaluation.message(
                 name, "strse", Integer(1), Expression(name, string, patt)
             )
-        return Expression("List", *[_search(re_patts, s, flags, matched) for s in py_s])
+        return Expression(SymbolList, *[_search(re_patts, s, flags, matched) for s in py_s])
     else:
         py_s = string.get_string_value()
         if py_s is None: